# Real-time-team-task-management/swagger.yaml
openapi: 3.0.3
info:
  title: Real-Time Team Task Management API
  version: "1.0.0"
  description: API for Real-Time Team Task Management system (Auth, Teams, Projects, Tasks, Socket.IO)

servers:
  - url: http://localhost:5000
    description: Local dev

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [admin, member] }
    AuthResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        token: { type: string }

    Team:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        description: { type: string }
        code: { type: string }
        members:
          type: array
          items:
            type: object
            properties:
              user: { $ref: '#/components/schemas/User' }
              role: { type: string }

    Project:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        description: { type: string }
        team: { type: string }

    Task:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        description: { type: string }
        priority: { type: string, enum: [low, medium, high] }
        dueDate: { type: string, format: date-time }
        status: { type: string, enum: [To Do, In Progress, Review, Done] }
        project: { type: string }
        assignee: { $ref: '#/components/schemas/User' }
        createdBy: { $ref: '#/components/schemas/User' }

security:
  - bearerAuth: []

paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
                role: { type: string, enum: [admin, member] }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }

  /api/team/create:
    post:
      tags: [Team]
      summary: Create team (admin only)
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        '201':
          description: Created team
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data: { $ref: '#/components/schemas/Team' }

  /api/team/my:
    get:
      tags: [Team]
      summary: Get teams current user belongs to
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  count: { type: integer }
                  teams:
                    type: array
                    items: { $ref: '#/components/schemas/Team' }

  /api/team/join/{teamCode}:
    post:
      tags: [Team]
      summary: Join team by team code
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: teamCode
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Joined
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  team: { $ref: '#/components/schemas/Team' }

  /api/project:
    post:
      tags: [Project]
      summary: Create a project under a team
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                team: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  project: { $ref: '#/components/schemas/Project' }

  /api/project/team/{teamId}:
    get:
      tags: [Project]
      summary: Get projects by team
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: teamId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  projects:
                    type: array
                    items: { $ref: '#/components/schemas/Project' }

  /api/task:
    post:
      tags: [Task]
      summary: Create task (authenticated)
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                priority: { type: string, enum: [low, medium, high] }
                dueDate: { type: string, format: date-time }
                project: { type: string }
                assignee: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  task: { $ref: '#/components/schemas/Task' }

  /api/task/project/{projectId}:
    get:
      tags: [Task]
      summary: Get tasks by project
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: projectId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  tasks:
                    type: array
                    items: { $ref: '#/components/schemas/Task' }

  /api/task/{id}:
    put:
      tags: [Task]
      summary: Update task (admin)
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Updated
    delete:
      tags: [Task]
      summary: Delete task (admin)
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted